+++ src/ast/ty.rs
@@ -5,7 +5,7 @@ pub type Type = Box<Type_>;
 
 #[derive(Clone, Debug)]
 pub struct Type_ {
-    pub pos: u32,
+    pub pos: usize,
     pub data: TypeData,
 }
 
@@ -21,37 +21,37 @@ pub enum TypeData {
 }
 
 impl Type_ {
-    pub fn void_type(pos: u32) -> Type {
+    pub fn void_type(pos: usize) -> Type {
         Box::new(Type_ {
             pos,
             data: TypeData::Void,
         })
     }
-    pub fn name_type(pos: u32, name: String) -> Type {
+    pub fn name_type(pos: usize, name: String) -> Type {
         Box::new(Type_ {
             pos,
             data: TypeData::Name(name),
         })
     }
-    pub fn array_type(pos: u32, ty: Type, size: Exp) -> Type {
+    pub fn array_type(pos: usize, ty: Type, size: Exp) -> Type {
         Box::new(Type_ {
             pos,
             data: TypeData::Array(ty, size),
         })
     }
-    pub fn pointer_type(pos: u32, ty: Type) -> Type {
+    pub fn pointer_type(pos: usize, ty: Type) -> Type {
         Box::new(Type_ {
             pos,
             data: TypeData::Pointer(ty),
         })
     }
-    pub fn template_type(pos: u32, name: String, ty_params: Vec<Type>) -> Type {
+    pub fn template_type(pos: usize, name: String, ty_params: Vec<Type>) -> Type {
         Box::new(Type_ {
             pos,
             data: TypeData::Template(name, ty_params),
         })
     }
-    pub fn func_type(pos: u32, params: field::FieldList, result: Type) -> Type {
+    pub fn func_type(pos: usize, params: field::FieldList, result: Type) -> Type {
         Box::new(Type_ {
             pos,
             data: TypeData::Func(params, result),
